# Conflicts:
#	HSEPortal.Client/src/app/features/application/application-completed/application-completed.component.ts
#	HSEPortal.Client/src/app/features/returning-application/verify.component.ts

+++ HSEPortal.Client/src/app/features/application/accountable-person/check-answers/check-answers.component.ts
@@ -3,7 +3,7 @@ import { ActivatedRoute, ActivatedRouteSnapshot } from '@angular/router';
 import { PaymentDeclarationComponent } from 'src/app/features/application/payment/payment-declaration/payment-declaration.component';
 import { PaymentModule } from 'src/app/features/application/payment/payment.module';
 import { FieldValidations } from 'src/app/helpers/validators/fieldvalidations';
-import { AccountablePersonModel, ApplicationService, BuildingApplicationStage } from 'src/app/services/application.service';
+import { AccountablePersonModel, ApplicationService, BuildingApplicationStage, Status } from 'src/app/services/application.service';
 import { AccountabilityArea } from 'src/app/components/pap-accountability/pap-accountability.component';
 import { AccountabilityAreasHelper } from 'src/app/helpers/accountability-areas-helper';
 import { PageComponent } from 'src/app/helpers/page.component';
@@ -18,17 +18,19 @@ export class AccountablePersonCheckAnswersComponent extends PageComponent<void> 
   static title: string = "Check your answers for PAP and AP - Register a high-rise building - GOV.UK";
 
   checkAnswersArea = AccountabilityArea.CheckAnswers;
+  apToRemove?: AccountablePersonModel;
 
   aps: AccountablePersonModel[] = [];
   constructor(activatedRoute: ActivatedRoute) {
     super(activatedRoute);
-  } 
+  }
 
   hasIncompleteData = false;
 
   override onInit(applicationService: ApplicationService): void {
-    this.aps = this.applicationService.model.AccountablePersons;
-    this.updateAddAnotherVariable(this.applicationService.model.AccountablePersons);
+    this.aps = this.applicationService.currentVersion.AccountablePersons;
+    this.updateAddAnotherVariable(this.applicationService.currentVersion.AccountablePersons);
+    if (this.isRAinProgress()) this.titleService.setTitle("Check your answers about accountable persons - Register a high-rise building - GOV.UK");
   }
 
   override canAccess(applicationService: ApplicationService, routeSnapshot: ActivatedRouteSnapshot): boolean {
@@ -44,7 +46,7 @@ export class AccountablePersonCheckAnswersComponent extends PageComponent<void> 
         canContinue &&= FieldValidations.IsNotNullOrWhitespace(ap.OrganisationType);
         canContinue &&= FieldValidations.IsNotNullOrWhitespace(ap.OrganisationName);
 
-        if (index == 0) {
+        if (index == 0 && !this.applicationService.isChangeAmendmentInProgress) {
           canContinue &&= FieldValidations.IsNotNullOrWhitespace(ap.Role);
         }
 
@@ -73,16 +75,14 @@ export class AccountablePersonCheckAnswersComponent extends PageComponent<void> 
         }
 
       } else if (ap.Type == "individual") {
-        if (ap.IsPrincipal == "yes") {
-          canContinue &&= (ap.PapAddress ?? ap.Address) != null;
-        } else {
-          canContinue &&= ap.Address != null;
+        canContinue &&= (ap.PapAddress ?? ap.Address) != null;
+        if (ap.IsPrincipal != "yes") {
           canContinue &&= FieldValidations.IsNotNullOrWhitespace(ap.FirstName);
           canContinue &&= FieldValidations.IsNotNullOrWhitespace(ap.LastName);
           canContinue &&= FieldValidations.IsNotNullOrWhitespace(ap.PhoneNumber);
           canContinue &&= FieldValidations.IsNotNullOrWhitespace(ap.Email);
 
-          if (index == 0) {
+          if (index == 0 && !this.isRAinProgress()) {
             canContinue &&= ap.PapAddress != null;
           }
         }
@@ -93,16 +93,20 @@ export class AccountablePersonCheckAnswersComponent extends PageComponent<void> 
         canContinue &&= (ap.SectionsAccountability?.findIndex(x => (x.Accountability?.length ?? 0) > 0) ?? -1) > -1;
       }
     }
-    
-    canContinue &&= this.applicationService.model.Sections.filter(x => !x.Scope?.IsOutOfScope).every(section => AccountabilityAreasHelper.getNotAllocatedAreasOf(this.applicationService.model.AccountablePersons, this.applicationService.model.BuildingName!, section).length == 0);
+
+    canContinue &&= this.applicationService.currentVersion.Sections.filter(x => !x.Scope?.IsOutOfScope && x.Status != Status.Removed).every(section => AccountabilityAreasHelper.getNotAllocatedAreasOf(this.applicationService.currentVersion.AccountablePersons, this.applicationService.model.BuildingName!, section).length == 0);
 
     this.hasIncompleteData = !canContinue;
     return canContinue;
   }
 
   override navigateNext(): Promise<boolean | void> {
+    if (this.applicationService.isChangeAmendmentInProgress) {
+      return this.navigationService.navigateRelative(`../registration-amendments/change-task-list`, this.activatedRoute);
+    }
+
     this.applicationService.model.ApplicationStatus = this.applicationService.model.ApplicationStatus | BuildingApplicationStage.AccountablePersonsComplete;
-    
+
     this.applicationService.updateApplication();
 
     if ((this.applicationService.model.ApplicationStatus & BuildingApplicationStage.PaymentComplete) == BuildingApplicationStage.PaymentComplete) {
@@ -112,8 +116,28 @@ export class AccountablePersonCheckAnswersComponent extends PageComponent<void> 
     return this.navigationService.navigateRelative(`../${PaymentModule.baseRoute}/${PaymentDeclarationComponent.route}`, this.activatedRoute);
   }
 
+  saveAndContinueOrKnockOnQuestions() {
+    var canContinue = true;
+    for (let index = 0; index < this.aps.length; index++) {
+      var ap = this.aps[index];
+      if (index > 0) {
+        canContinue &&= (ap.SectionsAccountability?.length ?? 0) > 0;
+        canContinue &&= (ap.SectionsAccountability?.findIndex(x => (x.Accountability?.length ?? 0) > 0) ?? -1) > -1;
+      }
+    }
+
+    canContinue &&= this.applicationService.currentVersion.Sections.filter(x => !x.Scope?.IsOutOfScope && x.Status != Status.Removed).every(section => AccountabilityAreasHelper.getNotAllocatedAreasOf(this.applicationService.currentVersion.AccountablePersons, this.applicationService.model.BuildingName!, section).length == 0);
+
+    if (!canContinue) this.navigationService.navigateRelative("areas-accountability", this.activatedRoute);
+    else this.saveAndContinue();
+  }
+
   override async onSave(): Promise<void> {
     await this.applicationService.syncAccountablePersons();
+
+    if (this.applicationService.isChangeAmendmentInProgress) {
+      this.applicationService.currentVersion.ApChangesStatus = Status.ChangesComplete;
+    }
   }
 
   navigateTo(url: string, apIndex: number) {
@@ -122,15 +146,61 @@ export class AccountablePersonCheckAnswersComponent extends PageComponent<void> 
     });
   }
 
+  navigateToSamePAPscreen() {
+    return this.navigationService.navigateRelative("../registration-amendments/same-pap", this.activatedRoute);
+  }
+
+  addAnotherAP() {
+    return this.navigationService.navigateRelative('add-more', this.activatedRoute);
+  }
+
   removeAp(ap: AccountablePersonModel, index: number) {
-    this.applicationService.removeAp(index);
-    this.updateAddAnotherVariable(this.applicationService.model.AccountablePersons);
+    if (this.applicationService.isChangeAmendmentInProgress) {
+      this.apToRemove = ap;
+    } else {
+      this.applicationService.removeAp(index);
+      this.updateAddAnotherVariable(this.applicationService.currentVersion.AccountablePersons);
+    }
+  }
+
+  removeConfirmAp(confirm: boolean) {
+    if (confirm) {
+      this.applicationService.removeAp(this.applicationService.currentVersion.AccountablePersons.indexOf(this.apToRemove!));
+      this.updateAddAnotherVariable(this.applicationService.currentVersion.AccountablePersons);
+    }
+
+    this.apToRemove = undefined;
   }
 
   private updateAddAnotherVariable(aps: AccountablePersonModel[]) {
-    if(!!aps && aps.length > 1) {
+    if (!!aps && aps.length > 1) {
       aps.slice(0, aps.length - 2).map(x => x.AddAnother = 'yes');
-      if(!!aps.at(-1)) aps.at(-1)!.AddAnother = 'no';
+      if (!!aps.at(-1)) aps.at(-1)!.AddAnother = 'no';
     }
+  }
+
+  isRAinProgress(): any {
+    return this.applicationService.isChangeAmendmentInProgress;
+  }
+
+  get previousPAPName() {
+    return this.getPAPName(this.applicationService.previousVersion.AccountablePersons[0]);
+  }
+
+  private getPAPName(pap: AccountablePersonModel) {
+    let individualName = pap.IsPrincipal == 'yes' && !FieldValidations.IsNotNullOrWhitespace(pap.FirstName) ? `${this.applicationService.model.ContactFirstName} ${this.applicationService.model.ContactLastName}` : `${pap.FirstName} ${pap.LastName}`;
+    return pap.Type == 'organisation' ? pap.OrganisationName : individualName;
+  }
+
+  get isStillPapSentence() {
+    return `Is ${this.previousPAPName} still the principal accountable person?`;
+  }
+
+  get nolongerAccountableSentence() {
+    return `${this.previousPAPName} areas of accountability`;
+  }
+
+  get newPap() {
+    return this.applicationService.model.RegistrationAmendmentsModel?.AccountablePersonStatus?.NewPap;
   }
 }
